11
1

+++ pyHelper/OStorage/COSBrowser.py
@@ -26,7 +26,6 @@ rootPath = os.path.split(curPath)[0]
 sys.path.append(rootPath)
 from qcloud_cos import CosConfig, CosS3Client
 import logging
-import re
 import yy_utils
 
 logging.basicConfig(level=logging.INFO, stream=sys.stdout)
@@ -35,10 +34,6 @@ upload_list_path = 'out/upload_list.txt'
 delete_list_path = 'out/delete_list.txt'
 test_bucket_bat = 'bucket_test_bg.bat'
 main_bucket_bat = 'bucket_main.bat'
-
-local_pre_path = 'E:/cache/root/'
-sync_local_dir = 'ttt'
-bucket_dir = 'ttt'
 
 
 # 同步操作.分3个步骤
@@ -52,13 +47,15 @@ bucket_dir = 'ttt'
 
 
 # ----------------功能1--------------------
-def local_list(pre_path, sync_local_path):
+def local_list(path):
     list = []
-    for root, dirs, files in os.walk(pre_path + sync_local_path):
+    for root, dirs, files in os.walk(path):
         for file in files:
+            if '.DS_Store' in file:
+                continue
             temp = os.path.join(root, file).replace('\\', '/')
             size = os.path.getsize(temp)
-            list.append('%s|%d' % (temp.replace(pre_path, ''), size))
+            list.append('%s|%d' % (temp.replace(path, ''), size))
     return list
 
 
@@ -66,9 +63,9 @@ def org_os_list(cmd_res):
     org = []
     for i in cmd_res:
         i = i.strip()
-        span = re.match(r'(.+) +(\d+) +\d{4}-\d{2}-\d{2}', i)
-        key = span.group(1).strip()
-        size = span.group(2).strip()
+        span = i.split()
+        key = span[0].replace(bucket_dir, '').strip()
+        size = span[1].strip()
         item = key + '|' + size
         org.append(item)
     return org
@@ -77,20 +74,20 @@ def org_os_list(cmd_res):
 def list_finish(res, _):
     logging.info('list os files suc! do next:save different list to file! out/upload_list.txt & out/delete_list.txt')
     os_list = org_os_list(res)
-    local_files = local_list(local_pre_path, sync_local_dir)
+    local_files = local_list(local_path)
     upload_list, delete_list = diff_path(os_list, local_files)
     upload_path = []
     for upload_item in upload_list:
-        upload_path.append(upload_item.split('|')[0])
+        upload_path.append(local_path + upload_item.split('|')[0])
 
     delete_px_list = []
     for delete_item in delete_list:
         if delete_item.split('|')[0] not in upload_path:
-            delete_px_list.append(delete_item)
+            delete_px_list.append(bucket_dir + delete_item.split('|')[0])
 
     yy_utils.create_dirs(upload_list_path)
     with open(upload_list_path, 'w', encoding='utf-8') as f:
-        f.write('\n'.join(upload_list))
+        f.write('\n'.join(upload_path))
 
     yy_utils.create_dirs(delete_list_path)
     with open(delete_list_path, 'w', encoding='utf-8') as f:
@@ -124,7 +121,7 @@ def create_diff_list(bat):
 # ---------------功能2---------------------------
 def sync_path(_, param):
     # 上传列表需要优化.一两张不需要处理成这样吧? 2000+文件上传需要几分钟 1w? 10分钟?
-    yy_utils.process_cmd('coscmd upload -rs %s %s' % (local_pre_path + sync_local_dir, bucket_dir))
+    yy_utils.process_cmd('coscmd upload -rs %s %s' % (local_path, bucket_dir))
     pass
 
 
@@ -143,9 +140,13 @@ def sync_to_os(bat):
         bucket = cmd_splite[bucket_index]
         config2 = CosConfig(Region=region, SecretId=secret_id, SecretKey=secret_key, Token=None, Scheme='https')
         client = CosS3Client(config2)
-    if os.path.exists(upload_list_path):
+    with open(upload_list_path, 'r') as f:
+        can_upload = len(f.readlines()) > 0
+    if can_upload and os.path.exists(upload_list_path):
         yy_utils.process_cmd(bat, done_call=sync_path)
-        pass
+    else:
+        print('没有要上传的文件!')
+
     if os.path.exists(delete_list_path):
         # 执行批量删除
         delete_list_obj = []
@@ -154,12 +155,15 @@ def sync_to_os(bat):
             for line in delete_list:
                 delete_list_obj.append({'Key': line.split('|')[0]})
             print(delete_list_obj)
-        client.delete_objects(
-            Bucket=bucket,
-            Delete={
-                'Object': delete_list_obj
-            }
-        )
+        if len(delete_list_obj) > 0:
+            client.delete_objects(
+                Bucket=bucket,
+                Delete={
+                    'Object': delete_list_obj
+                }
+            )
+        else:
+            print('没有要删除的文件!')
 
 
 # ---------------功能2结束------------------------
@@ -176,24 +180,21 @@ if __name__ == '__main__':
             # print('o:' + o + '  a:' + a)
             if a.endswith('/') or a.endswith('\\'):
                 a = a[:-1]
-            local_pre_path = os.path.dirname(a) + '/'
-            sync_local_dir = os.path.basename(a)
+            local_path = a
         if o in ('-b', '--bucket'):
             print(a)
             if a.endswith('/') or a.endswith('\\'):
                 a = a[:-1]
             bucket_dir = a
-    bucket_dir = 'media/pic'
-    sync_local_dir = r'pic'
-    local_pre_path = r'E:/resource/desc/'
-    print(local_pre_path, sync_local_dir, bucket_dir)
+    bucket_dir = 'media/pic/thum'
+    local_path = r'/Users/mryang/Documents/res/src/pic/thum'
+    print(local_path, bucket_dir)
     # 这是两步操作,通常需要分开
     create_diff_list(main_bucket_bat)
-    # os.system(upload_list_path)
     os.system('open ' + upload_list_path)
     os.system('open ' + delete_list_path)
     input = input('去确认上传和下载文件吧!:(y|n)')
-    if (input == 'y'):
+    if input == 'y':
         print('确认')
         sync_to_os(main_bucket_bat)
     else:
